+++ R/blob_transfer_internal.R
@@ -13,17 +13,22 @@ upload_blob_internal <- function(container, src, dest, type, blocksize, lease=NU
 }
 
 
-download_blob_internal <- function(container, src, dest, blocksize=2^24, overwrite=FALSE, lease=NULL, check_md5=FALSE)
+download_blob_internal <- function(container, src, dest, blocksize=2^24, overwrite=FALSE, lease=NULL,
+                                   check_md5=FALSE, snapshot=NULL)
 {
     headers <- list()
     if(!is.null(lease))
         headers[["x-ms-lease-id"]] <- as.character(lease)
 
+    opts <- if(is.null(snapshot))
+        list()
+    else list(snapshot=snapshot)
+
     dest <- init_download_dest(dest, overwrite)
     on.exit(dispose_download_dest(dest))
 
     # get file size (for progress bar) and MD5 hash
-    props <- get_storage_properties(container, src)
+    props <- get_storage_properties(container, src, snapshot=snapshot)
     size <- as.numeric(props[["content-length"]])
     src_md5 <- props[["content-md5"]]
 
@@ -33,7 +38,8 @@ download_blob_internal <- function(container, src, dest, blocksize=2^24, overwri
     while(offset < size)
     {
         headers$Range <- sprintf("bytes=%.0f-%.0f", offset, offset + blocksize - 1)
-        res <- do_container_op(container, src, headers=headers, progress=bar$update(), http_status_handler="pass")
+        res <- do_container_op(container, src, options=opts, headers=headers, progress=bar$update(),
+                               http_status_handler="pass")
         httr::stop_for_status(res, storage_error_message(res))
         writeBin(httr::content(res, as="raw"), dest)