+++ R/storage_metadata.R
@@ -3,6 +3,7 @@
 #' @param object A blob container, file share or ADLS filesystem object.
 #' @param blob,file Optionally the name of an individual blob, file or directory within a container.
 #' @param isdir For the file share method, whether the `file` argument is a file or directory. If omitted, `get_storage_metadata` will auto-detect the type; however this can be slow, so supply this argument if possible.
+#' @param snapshot For the blob method of `get_storage_metadata`, an optional snapshot identifier. This should be a datetime string, in the format "yyyy-mm-ddTHH:MM:SS.SSSSSSSZ". Ignored if `blob` is omitted.
 #' @param ... For the metadata setters, name-value pairs to set as metadata for a blob or file.
 #' @param keep_existing For the metadata setters, whether to retain existing metadata information.
 #' @details
@@ -47,14 +48,19 @@ get_storage_metadata <- function(object, ...)
 
 #' @rdname metadata
 #' @export
-get_storage_metadata.blob_container <- function(object, blob, ...)
+get_storage_metadata.blob_container <- function(object, blob, snapshot=NULL, ...)
 {
     if(missing(blob))
     {
         options <- list(restype="container", comp="metadata")
         blob <- ""
     }
-    else options <- list(comp="metadata")
+    else
+    {
+        options <- list(comp="metadata")
+        if(!is.null(snapshot))
+            options$snapshot <- snapshot
+    }
 
     res <- do_container_op(object, blob, options=options, http_verb="HEAD")
     get_classic_metadata_headers(res)